Cleaned Code, Refactored UI to UI Class

+++ OrbitSim/src/Orbit.h
@@ -7,12 +7,12 @@
 //#include <vector>
 
 // UNIVERSAL CONSTANTS
-#define PI 3.14159265
+#define PI static_cast<float>(3.14159265)
 const float G_M = 4 * PI * PI; // G * sunMass constant
-const long float sunMass = 1.989e30;
+const float sunMass = 1.989e30;
 // Force Constant predicted by General Relativity
-#define ALPHA_FACTOR 1.1e-8 // AU^2 - correction factor as predected by general relativity
-#define FCONST(m,r,B) ( (G_M * m) / pow(r, B + 1.0f) ) * (1+ (ALPHA_FACTOR/pow(r,2)))
+#define ALPHA_FACTOR static_cast<float>(1.1e-8) // AU^2 - correction factor as predected by general relativity
+#define FCONST(m,r,B) static_cast<float>(( (G_M * m) / pow(r, B + 1.0f) ) * (1+ (ALPHA_FACTOR/pow(r,2))))
 
 class Orbit
 {
@@ -51,7 +51,7 @@ public:
     float t;
     float dt;
 
-    // Statistics
+    // Orbit Statistics
     float aphelion;   // max radius
     float perihelion; // min radius
     bool aphelionReached;
